Aven-backend/requirements.txt
flask
pymongo
pinecone-client
openai
google-generativeai
requests